6gn-functions/release/requirements.txt
paho-mqtt
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
grpcio